Only flush extents that might actually be dirty. (#875)
This change keeps track at the region level which extents might be
dirty. That way, during a flush, it only locks/sends flush commands to
those extents.

This makes a big difference for regions with a lot of extents. Right now
our write perf goes down the more extents you have, and it's pretty
gnarly.

When opening extents, we check the dirty flag. If it's dirty, we add it
to the region's dirty flag cache.

When we send a write command to an extent, we unconditionally set the
dirty flag cache for that extent. If the extent decides the write is a
NOP and doesn't actually mark itself dirty in its metadata, it'll still
get flushed. That's fine- it's technically a waste to do, but it's not
going to happen a lot and it's minor if it does. So it's not worth the
complexity we'd need to perfectly track the dirty state of the extents.

Basically, an incorrect negative is unacceptable, but an incorrect
positive is fine.