Test update for fast write ack and gather errors on test failure
Updated a few tests, including tools/test_up.sh  to accept a range of
valid write counters where a write after a flush might not make it to
disk before the test stops.  This fixes issues #896

Added some additional code to test-live-repair buildomat job to gather info
in the event of failure.