Only rehash if a write may have failed (#899)
Previously, we always rehashed data during a `flush` operation, which required
reading it back from disk and computing an integrity hash.

However, if a block was successfully written, then we already know what its hash
value will be (since we computed it earlier)!  We _only_ need to read + rehash
if we failed to write the data, because if we failed to write the data, then the
on-disk bytes are in an unknown state.

This PR tracks dirty blocks as a map from `block -> Option<u64>`, so that if the
hash is known, it won't be recomputed during a flush operation.

It shows a small but noticeable (1-3%) performance increase when running
`pgbench`.