Crutest updates, rename tests, new options
Renamed the Rand test to WFR, as it actually was doing a write then
a flush then a read and that better describes what it does.

Renamed verify option to verify-at-end

Created a new verify-at-start and made that option require the
verify-in option through clap instead of a specific check.