Don't pass MutexGuard into functions (#917)
Small cleanup: `MutexGuard` implements `deref_mut` into a plain mutable
reference, so we can use that instead.

Note that this could _theoretically_ change behavior in `submit_flush_internal`,
which took the `MutexGuard` by value (i.e. dropping it at the end of the
function).

Now, the `MutexGuard` is now dropped _outside_ of `submit_flush_internal`.
However, the call to `submit_flush_internal` is the last call in all parent
function, so there's no actual behavior change.

All other functions changed in this PR were taking a `&mut MutexGuard`, which is
harmless to change to a `&mut` reference.