Fast return for zero length IOs
If a read or write with length zero is submitted to either
the Volume layer or via a Guest, we just return success without
having to involve any of the layers below us.

Updated crutest to allow a zero length IO.
Added tests to verify a zero length IO via Volume or Guest
does not return error or panic.