Use a strong type for client ID (#925)
Crucible always runs with three clients, and we need a "client ID" data type.

Previously, client ID was (usually) a `u8`, and "data `T` for every client" was
variously

- `Vec<T>` (with 3 elements, guaranteed)
- `Vec<T>` (_usually_ with 3 elements, but with 0 elements during some tests)
- `[T; 3]` (the only type which _actually_ guarantees 3 elements)
- `HashMap<u8, T>` (with 3 elements, guaranteed)
- `HashMap<u8, T>` (with 0-3 elements)

This PR has two goals:
- Replace the `u8` with a `ClientId(u8)` (guaranteed at construction to be
  `0..3`), and fix everything that broke when I did that.  This is similar to
  #919, which did the same thing for job ID.
- Replace the eclectic set of data structures above with only two things:
    - `ClientData<T>`, which represents "3 values, must be present"
    - `ClientMap<T>`, which is equivalent to a `HashMap<ClientId, T>` but is
      specialized to only contain 3 values max.  (Under the hood, this is
      implemented as `ClientData<Option<T>>`)

Both of these new data structures can be indexed with a `ClientId` directly, so
this removes a vast number of `client_id as usize` casts.  The new data
structures are free of allocation, require zero hashing, and can be copied
instead of cloned (if `T` is `Copy`).

`ClientId` is annotated with `#[serde(transparent)]`, so it looks the same on
the wire and should work fine with mixed upstairs / downstairs versions.