Keep new jobs sorted (#929)
When we start hitting flow control, new jobs are repeatedly removed from
`ds_active`, sorted, then inserted back using `Vec::extend_from_slice` (because
we hit the queue limit).  This shows up in the profiling data:

There are two ways to make this cheaper:
- Keep them sorted by construction (using a `BTreeSet` instead of a `Vec`)
- Use `append` instead of `extend`.  There are special cases in
  `BTreeSet::append` if one or other is empty, so `enqueue_work` will be more
  efficient if no one else added work to the queue in the meantime (and should
  be equivalent cost otherwise).

This is an incremental speedup for `test_successful_live_repair`.  It brings it
from 17.25 → 14.9 sec on my machine.