Wait for a snapshot to finish on all downstairs (#920)
Nexus, when running a snapshot-create saga, will attempt to start
read-only downstairs corresponding to a snapshot after it has returned
from taking that snapshot (either through the instance's propolis or via
the pantry), but the upstairs acks all non-deactivate flushes after only
two jobs have been seen. This can lead to a 404 trying to start a
read-only downstairs (as the snapshot may not exist yet). See #431.

Wait for all downstairs to return Ok for flush + snapshot jobs before
acking to the Guest.