Use `cargo hakari` for better workspace deps (#956)
Right now, if you build different packages in the Crucible workspace, they
enable different sets of features for downstream dependencies.  This means that
switching packages may require recompiling many dependencies (with slightly
different features), which slows down iteration.

[`cargo hakari`](https://docs.rs/cargo-hakari/latest/cargo_hakari/index.html) is
a tool to improve this workflow by ensuring that all packages in the workspace
enable the same (maximal) set of features.

This PR sets it up for the Crucible workspace, and updates the README
accordingly.