Fast dependency checking (#916)
# The problem of dependency tracking
Crucible jobs may be delivered to downstairs instances out of order.  Naively,
this is bad news: reads and writes can't be executed in arbitrary order and get
consistent results.  To avoid this issue, each job includes a list of
dependencies, which are jobs that must be finished before that job is executed.

Dependencies are tracked a per-block basis.  Right now, they are calculated by
iterating backwards over every single active job and picking out the ones with
overlapping affected blocks; flushes count as affecting every block.

This is not particularly efficient, for two different reasons:
- We have to iterate over a lot of jobs.  Our longest-running tests (e.g.
  `test_successful_live_repair`) are spending the **vast majority** of their
  time iterating over jobs repeatedly; this is a classic example of
  [Accidentally Quadratic](https://accidentallyquadratic.tumblr.com/) behavior.
- We send a bunch of unnecessary dependencies.  Here's an example from one of
  our repair tests:
```
        //     |       | UNDER |       |
        //     |       | REPAIR|       |
        //     | block | block | block |
        // op# | 0 1 2 | 3 4 5 | 6 7 8 | deps
        // ----|-------|-------|-------|-----
        //   0 |       |       | RpRpRp|
        //   1 |       |       | RpRpRp| 0
        //   2 |       |       | RpRpRp| 0,1
        //   3 |       |       | RpRpRp| 0,1,2
        //   4 | W W W | W W W | W W W | 0,1,2,3
        //   5 | R R R | R R R | R R R | 0,1,2,3,4
        //   6 | WuWuWu| WuWuWu| WuWuWu| 0,1,2,3,4,5
```

Eyeballing this example, we could reduce it to the following without loss of
ordering:
```
        //     |       | UNDER |       |
        //     |       | REPAIR|       |
        //     | block | block | block |
        // op# | 0 1 2 | 3 4 5 | 6 7 8 | deps
        // ----|-------|-------|-------|-----
        //   0 |       |       | RpRpRp|
        //   1 |       |       | RpRpRp| 0
        //   2 |       |       | RpRpRp| 1
        //   3 |       |       | RpRpRp| 2
        //   4 | W W W | W W W | W W W | 3
        //   5 | R R R | R R R | R R R | 4
        //   6 | WuWuWu| WuWuWu| WuWuWu| 5
```

# So, how do we fix this?

The answer is a new data structure which maintains a map from `ImpactedBlock` to
job dependencies.  You can think of this as tracking a wavefrom of active jobs
on a per-block basis; when a new job comes it, we
- Get its dependencies by checking its `ImpactedBlock` range
- Update the map, setting this job as the most recent dependency for its
  impacted blocks

We wouldn't want to maintain this map on a per-block basis; reads and writes often span many blocks, so iterating over every single block would be inefficient.  Instead the map stores dependencies over **ranges** of blocks (`std::ops::Range<ImpactedAddr>`).  Then, it's a small matter of programming to correctly split / merge / iterate over block ranges when handling dependencies.

This is a tricky data structure, and is tested using `proptest`.

# Reads versus writes
It's slightly more subtle than storing a _single_ active job for each block
range, because read jobs don't interfere with each other (and we therefore don't
want to impose an ordering on them).  Instead, we store both
- The last blocking job (`Option<u64>`); this is a write / flush / repair /
  anything but read
- All non-blocking jobs (`Vec<u64>`); these are reads

A new blocking job clears the non-blocking list and replaces the blocking job; a
new non-blocking job is appended to the non-blocking list but does not modify
the last blocking job.

# Live repair dependencies
Live repairs are a tricky special case: if a write operation spans the
live-repair extent, we would reserve new job IDs for the live repair, adding
them to `repair_job_ids` but not to `ds_active` (and then jumping through a
bunch of hoops to make sure that we respect their dependencies).

Tracking jobs on a per-block basis makes it easier: when we need to reserve live
repair job IDs, we get _their_ dependencies with `deps_for_live_repair`, leaving
the `ExtentLiveReopen` job ID as the active one for that whole extent.  Then,
later jobs (i.e. the read or write) which use that extent will depend on the
live repair job _automatically_. This dramatically simplifies
`submit_read/write/flush`.

In general, the new code is more homogeneous between normal operations and
reserved live repair jobs.  The pattern is as follows:
- Get the job ID (using `next_id` or similar)
- Compute the job's dependencies and install the job ID into the map, using
  `deps_for_read/write/flush`
- Enqueue the job once you've got a `DownstairsIO`.  For normal operations, this
  happens immediately; for reserved repair jobs, it happens when the later
  extent is reached by live repair.

# Other changes:
- Use `u64` for extent IDs (almost) everywhere, instead of a mix of `u64`,
  `u32`, and `usize`
- Tons of unit test fixing, because many dependencies changed!
- It's now harder to forge job IDs in unit tests, because we expect any job ID
  that's enqueued to _previously_ be installed into the dependency tracking map.
  Assertions will fire if you forget this, so it's hard to get wrong, but
  required a bunch of annoying plumbing.