The Pantry should reject non-block sized writes (#1013)
Instead of panicking, the Pantry should reject non-block sized reads and
writes with a 400 level response. The panic came from Volume, which has
been corrected to return a CrucibleError instead.

Unfortunately everything the Pantry returns is currently a 500: this
commit also defines different HttpError conversions for CrucibleError,
and the Pantry now returns those.

This commit also removes the `impl Into<std::io::Error> for
CrucibleError`: the author of that code didn't realize that `impl From`
works as long as one of the types is in the current crate!