Add raw file backend (#991)
This PR adds a new backend for Crucible Downstairs extent files: a single
on-disk file.  The on-disk file includes the raw block contents, a small
metadata block, and two slots per block for storing encryption context.

By alternating between slots, we get the same crash consistency behavior as the
SQLite database, but with many fewer syncs – we only need to add a `fsync` call
if we're about to overwrite a context slot that hasn't yet been synched to disk,
i.e. someone has written the same block 3x in a row without a flush.

The new backend implements the same `trait ExtentInner` as the SQLite backend,
introduced in #982 .  This means that it's _mostly_ seamless, although I had to
tweak a few small things around the edges.

Though this is a change to the on-disk format, we maintain support for existing
(SQLite-based) volumes.  Existing volumes fall into two categories, which we
handle differently:

- Snapshots are stored on read-only filesystems, so we can't mess with them.  We
  keep the old `SqliteInner` implementation around and use it in this case.
- Volumes that _aren't_ read-only are migrated from SQLite to raw extent files
  when they are opened.  The migration is done as an append to the existing data
  file, using the existance of the `.db` file as a marker to whether the
  migration is complete.

To simplify the implementation, this change requires a whole-rack update.
Specifically, live repair between SQLite extents and raw file extents is not
possible!  We have bumped the `CRUCIBLE_MESSAGE_VERSION` to prevent the system
from trying such a thing (though the message format remains identical).

One subtle point: we still have to support `.replace` directories with SQLite
files, because a live repair of an older extent could have been interrupted
before the system update!