Avoid unnecessary allocations during frame decode
Implementations of Decoder::decode should ensure that the buffer has
appropriate capacity for the rest of the bytes of a frame. Our code was
previously reserving the frame's total length, not space for the rest of
the frame we're currently reading. This commit fixes that to avoid
allocating unnecessary space.

Also, bail out during Encoder::encode if the frame is larger than
MAX_FRM_LEN: the decode step will fail anyway.