Remove Arc + Mutex from Buffer
By transfering Buffer's ownership to the BlockIO implementation instead
of passing by reference, the Arc + Mutex that guards Buffer's internals
can be removed. Return the Buffer when done so that ownership is
transferred back to the caller. Callers will have to change like so:

    -        crucible.read(offset, buffer.clone()).await?;
    +        let buffer = crucible.read(offset, buffer).await?;

Non-read operations are not affected.

Additionally, this commit removes the ability for one Guest operation to
be composed of multiple Downstairs operations: this support could be
added back, but enough was changing in this commit as is :) We don't
currently use this "bulk" functionality except in tests.