Offload write encryption (#1066)
This PR moves write encryption out of the main task.

Doing so is harder than it sounds, because we want to **preserve ordering** of
incoming `BlockReq`, but only `BlockReq::Write` and `BlockReq::WriteUnwritten`
need this encryption offload.

I turned to `FuturesOrdered`, which allows you to store a set of futures which
resolve in FIFO order.  For most `BlockReq`, we await a
`futures::future::ready(..)`, which simply provides the same `BlockReq` back to
us.  For writes, we offload the work using `rayon::spawn`, then return the
result through a `tokio::sync::oneshot` channel.  This strategy is recommended
by [this blog post by a Tokio maintainer](https://ryhl.io/blog/async-what-is-blocking/).

We can then store `Either<Ready, oneshot::Receiver>` in a `FuturesOrdered`,
which works out nicely.

The PR introduces a new generic `DeferredQueue` to organize these futures.  (The
`DeferredQueue` will also be used to defer read decryption in a subsequent PR).
When a `BlockReq` comes in, we put it into this queue*; futures being available
on the queue are another event source for the `select / apply` loop.

This is a significant speedup for large random writes:

| IO               | `main`     | `offload-write-encryption` |
|------------------|------------|----------------------------|
| 4K random writes | 27.2 MiB/s | 26.7 MiB/s                 |
| 1M random writes | 609 MiB/s  | 710 MiB/s                  |
| 4M random writes | 629 MiB/s  | 890 MiB/s                  |

I'm not sure if the performance delta for 4K writes is significant; if folks are
worried, we could special-case sufficiently small writes to be executed in the
main task.

*as an optimization, we execute non-write `BlockOp`s immediately if there is
nothing in the queue, to avoid an extra event