Comment cleanup and rename of DsState::Repair -> Reconcile
Some misc comment cleanup and a rename of the Downstairs state
Repair to Reconcile.  The term "repair" was often used in both
the reconcile case (what happens at start up) and during live
repair (what happens when the upstairs is active).  This cleans
up the comments and renames the original DsState::Repair to be
called DsState::Reconcile.

There should be no functional change from this.