Remove Arc + Mutex from Buffer (#1094)
Remove the Arc + Mutex that guards Buffer's internals and pass Buffer by
mutable reference instead. Callers will have to change like so:

    -        crucible.read(offset, buffer.clone()).await?;
    +        crucible.read(offset, &mut buffer).await?;

Non-read operations are not affected. Callers can reuse the same Buffer
(and resize it with `reset`) and avoid re-allocating for every read.
This is taken advantage of in `Volume::read`.

Additionally, this commit removes the ability for one Guest operation to
be composed of multiple Downstairs operations: this support could be
added back, but enough was changing in this commit as is :) We don't
currently use this "bulk" functionality except in tests.