Remove a superfluous copy during write serialization (#1087)
Right now, we copy data from a `BlockOp::Write/WriteUnwritten` into a `BytesMut`
buffer, encrypt it in-place, then serialize it using `CrucibleEncoder`.
Serialization requires a *second* full copy of the data.  This is noticeable in
our flamegraphs for large writes; see the PR for images.

This change removes that `memcpy`, speeding up large writes by a noticeable
amount; when the PR was first opened, I saw a ~30% speedup for 1M and 4M random
writes.