HEY! LISTEN! HEY! LISTEN!
Send Nexus a notification for the start, completion (successful or with
an error), and progress for any live repair or reconcilation.

For any Downstairs client stop request, send Nexus a notification. Send
a separate notification if a Downstairs client stops for any reason.

Notably, send everything and let Nexus filter out what it considers
noise. It will still log the reception of that notification which could
be a useful breadcrumb during debugging.

This commit introduces the "omicron-build" feature, and gates the
notification routines behind that.