Remove unnecessary mutexes from Downstairs (#1132)
Right now, a job going through the Downstairs has to lock three async mutexes:

- The main `Arc<Mutex<Downstairs>>`
- For a specific extent, an `Arc<Mutex<ExtentState>>`
- Within that extent, a `Mutex<Box<dyn ExtentInner>>`

Only the first one is necessary: once we have exclusive ownership over the
`Downstairs`, no one else can mess with it, so the other mutexes aren't
protecting us from anything.  This PR removes the other two locks.

There's only one tricky part:

Because we hold a `&Extent` across yield points, it must be `Send` (and
therefore `Extent` must be `Sync`).  It turns out that `rustqlite::Connection`
is **not** `Sync`, because it uses a `RefCell` internally.  Previously, storing
the `dyn ExtentInner` in a `Mutex` let us get away with this.

I went for the simplest-possible fix: wrapping the entire `SqliteInner` in a
`std::sync::Mutex`.  This is still one fewer mutex lock than the previous code,
and it means we don't have to change any other code.