Verify extent under repair is valid after copying files
Added a call to the downstairs repair API that will verify that
an extent is closed (or the region is read only), meaning the extent
is okay to copy.

This prevents a rare but possible condition where an extent is closed
when a remote side contacts it for repair, but is re-opened before the
extent file copy starts and the resulting copy is invalid.