Drop jobs from Offline downstairs. (#1157)
* Drop jobs from Offline downstairs.

If we receive a job ACK from an offline downstairs, and that job
has already been moved back to New, then discard it as we know
the replay will resend it when the downstairs comes back.

Added a connection_id field to the Client. Make a ConnectionId type
Use the ConnectionId type.