Do not 500 on snapshot delete for deleted region
The code path that validates a snapshot delete request checks that a
region exists, and returns 404 if it does not:

    match rc.context().get(&p.id) {
        Some(_) => (),
        None => {
            return Err(HttpError::for_not_found(
                None,
                format!("region {:?} not found", p.id),
            ));
        }
    }

However there's a flaw in this check: if the region existed but was
deleted, this will not return a 404, and proceed into `delete_snapshot`.

`delete_snapshot` will check if a running snapshot exists, and if it
does will bail out: it's an invalid request to delete a snapshot for
which there's a read-only downstairs running for.

However, the very next thing the code does is try to make a ZFSDataset
object out of the region's data path. If the region was deleted this
will lead to `zfs list` exiting with a non-zero return code, and will
bubble up to the user as a 500.

Catch ZFSDataset returning Err, and if the region existed but was
deleted, we can assume the snapshot was deleted prior to this (as the
agent would not allow otherwise). If the region existed and isn't
deleted, and `zfs list` for that region path failed, then the agent's
datafile doesn't match reality.