Add generic 'message or raw message' code (#1182)
This PR adds a generic implementation of the 'Message or raw message' serializer
used for reduced-mempcy serialization (#1087), then updates `crucible-upstairs`
to use the new implementation.

It also more generally reduces allocations in the socket write path by reusing
the same `BytesMut` (for `Message` serialization) or a `Vec` (for raw message
headers).  This may have a small performance impact; I didn't measure it.

These changes are a building block for faster serialization of Downstairs
`ReadResponse`, but I pulled them into a standalone PR for ease of review.