Make three smaller timeouts instead of one large one.
Updated the Upstairs <-> Downstairs message timeout to have
three smaller timeouts before triggering a disconnect due to
inactivity.  Log a message on each smaller timeout.

Moved deadline_secs() to a common location for both upstairs
and downstairs to use.  This resulted in a little shuffling
of included libraries as well.