Make three smaller timeouts instead of one large one. (#1207)
Make three smaller timeouts instead of one large one.

Updated the Upstairs <-> Downstairs message timeout to have
three smaller timeouts before triggering a disconnect due to
inactivity.  Log a message on each smaller timeout.

Made a global verbose_timeout for up and down to use.

Updated upstairs and downstairs to use the new verbose_timeout.

To fix the lossy select path preventing the timeout select path from
ever being taken, I first hoisted the lossy check into do_work_task()
and select over it and receiving a message indicating new work.

However, this was a misguided effort.  Due to the way the do_work_for
loop now operates, we will always keep looping till all the work we
can do is done. This means we no longer needed the lossy in the select
to begin with!  The final change is just to remove it.

All of this has also shown that the lossy numbers need some serious
tuning to be of any value.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>