DTrace, cmon, cleanup.
The counters for reconcile_repair_needed and reconcile_repaired
got lost at some point in the past, this will put them back.
Added count of reconciliation aborted to the DTrace Arg.
Added the extent_limit to the set of values in the DTrace Arg.

Replaced an assert with an if/panic to print more information if
we do encounter the situation.

Gather and log a more accurate count of jobs that we moved to replay.

Print fewer messages during reconciliation.

Added a 10 second delay before we reconnect to a downstairs.

Overhauled the cmon program a bit.
Added more to the dtrace | cmon path, where we can select more of
the values in the DTrace Arg and print them out.
Added a short version of the downstairs state to allow for narrow terminal
printing.  Added extent_limit, backpressure, next_job_id, job_id_delta,
ds_delay, and an io_summary.