Return *410 Gone* if volume is inactive (#1232)
If the Pantry's volume is no longer active when it comes time to serve a
request, return `410 Gone` instead of 503, as 503 is interpreted as a
temporary problem. If the Pantry's volume is no longer active, then it's
likely a Propolis has activated and taken over, which should signal an
unwind of whatever saga is using the Pantry.