`crutest` cleaning; adding rand-read/write tests (#1233)
This is a grab-bag of `crutest` improvements and features, with the goal
of making benchmarking and performance evaluation easier.

# New features
Adding `rand-read` and `rand-write` workloads. This is meant to mimic
the `fio` workload, but running outside of the Propolis VM to remove one
layer of abstraction.

# Medium optimization
Refactor `fill_workload` and `verify_volume` to use 8x worker tasks and
run in parallel. This is a significant speedup: for a 32 GiB region,
running fill + verify goes from 7:47 down to 1:35.

# Minor tweaks
- Use `&Guest` instead of `Arc<&Guest>` in function arguments where
possible
- Use `default_value_t` instead of `default_value` so that default
arguments are checked at compile-time
- Make some functions take `&self` instead of `&mut self`
- Delete unused functions (`RegionInfo::set_wc_min`)
- Replace `Vec<Future<..>>` with `FuturesOrdered` in a few places, and
remove our custom `join_all` function