Dummy downstairs cleanup (#1247)
I want `dummy_downstairs.rs` to be easier to use, since it's a great way
to do end-to-end testing.

In the current code, the `FramedRead/Write` are shared by both the main
(test) task and IO worker task; this shared state makes the code harder
to follow, because there are `lock().await`s everywhere. The division of
labor is also unclear: in some cases, the test task will write to those
streams directly; in other cases, it will use `mpsc` channels to the
worker task.

This PR refactors the IO worker task to **own** the `FramedRead/Write`;
the test task now **always** communicates to it via `mpsc` channels, and
cannot read or write to the streams directly.

This removes a bunch of locks, for a significant LOC decrease.

It also removes `bail_assert!` and `bail_assert_eq!` in favor of their
normal counterparts.