Add Upstairs session ID to dtrace stat probe, cleanup closure (#1254)
The one task refactor opened the door to lockless stat gathering for the
upstairs. Since we no longer need to grap locks to gather this data, we
can stuff it all inside the closure. This should mean we only pay the stat
collecting tax when the probe is actually enabled.

Added another field to the stat probe that records the session ID. This
will be unique not only per PID, but also for each Upstairs that could be
running inside a single PID. This is specifically useful when propolis has more than
one virtual disk attached to the same instance.