Agent update to support cloning.
This updates the crucible agent with new fields for the agent's
Region structure as well as the CreateRegion structure.  These new
fields tell the agent that it should create a region by cloning
another region at the provided IP and Port, which involves a
slightly different set of steps during region create.

The Region source and read_only field are also persisted in the agents
database file and a region created from a source must be a read only
region.  Supporting a read only region requires updating the SMF
service to add a 'mode' option that is passed to the downstairs.

I also added a bunch of agent-antagonist updates to test the clone
behavior.  Now, on the random match, we will also:
- Start the snapshot running that we just created.
- create a new region using clone where the source is our new snapshot.
- delete the new clone region we just created.

In addition, agent was updated to listen for a SIGUSR1 signal, and when
it is received, we shutdown all the tasks cleanly.
Agent will now watch for any task to panic/stop and if detected, it will
stop the rest of the tasks.

Refactor the agent-antagonist into reusable functions.

Updated the agent endpoint with new fields and the ability to create
a region as a clone if a source endpoint is provided.

Some further thoughts:
Region clone should use a global repair offset instead of hard coded 4000

Do we want to add additional fields to the agents data file at this time?
Like perhaps a time stamp when updated, or a new format that will identify
which version we are running?

Can we make a test to update the crucible.json file from old to new?
Better names for the old structure version?