Agent update to support cloning. (#1262)
This updates the crucible agent with new fields for the agent's Region
structure as well as the CreateRegion structure. These new fields tell
the agent that it should create a region by cloning another region at
the provided IP and Port, which involves a slightly different set of
steps during region create.  The agent was updated to, when receiving
the new fields populated, to create the region with the clone option.

The Region source and read_only field are also persisted in the agents
database file and a region created from a source must be a read only
region. Supporting a read only region requires updating the SMF service
to add a 'mode' option that is passed to the downstairs.

I also added a bunch of agent-antagonist updates to test the clone
behavior including a large refactor so I could re-use parts of the code.
Now, on the random match, we will also:
- Start the snapshot running that we just created.
- create a new region using clone where the source is our new snapshot.
- delete the new clone region we just created.

In addition, agent-antagonist was updated to listen for a SIGUSR1 signal, and when
it is received, we shutdown all the tasks cleanly. Agent will now watch
for any task to panic/stop and if detected, it will stop the rest of the
tasks.
---------

Co-authored-by: Alan Hanson <alan@oxide.computer>