Remove ReqwestBlockIO and VolumeConstructionRequest::Url
As far as I can tell, nothing uses this anymore. With the exception of a
pantry integration test. That test is not interested in the Reqwest
part, and is just interested in a read-only parent with random data in
it, so I changed that to use a tempfile read-only parent.

Propolis will need to be updated. It has exactly one reference to
`VolumeConstructionRequest::Url`:

    bin/propolis-server/src/lib/initializer.rs
    385:                    VolumeConstructionRequest::Url { id, .. } => id.to_string(),

This is in some code that's extracting the UUIDs out of all the VCR
types, and also doesn't care about the Reqwest functionality.